Move request errors to its own file
